+++ plugins/linear-genome-view/src/LinearGenomeView/components/HeaderPanControls.tsx
@@ -1,0 +1,47 @@
+import ArrowBackIcon from '@mui/icons-material/ArrowBack'
+import ArrowForwardIcon from '@mui/icons-material/ArrowForward'
+import { Button, alpha } from '@mui/material'
+import { makeStyles } from 'tss-react/mui'
+
+import { SPACING } from '../consts'
+
+import type { LinearGenomeViewModel } from '..'
+
+type LGV = LinearGenomeViewModel
+const useStyles = makeStyles()(theme => ({
+  panButton: {
+    background: alpha(theme.palette.background.paper, 0.8),
+    color: theme.palette.text.primary,
+    margin: SPACING,
+  },
+
+  buttonSpacer: {
+    marginRight: theme.spacing(2),
+  },
+}))
+
+export default function HeaderPanControls({ model }: { model: LGV }) {
+  const { classes } = useStyles()
+  return (
+    <>
+      <Button
+        variant="outlined"
+        className={classes.panButton}
+        onClick={() => {
+          model.slide(-0.9)
+        }}
+      >
+        <ArrowBackIcon />
+      </Button>
+      <Button
+        variant="outlined"
+        className={classes.panButton}
+        onClick={() => {
+          model.slide(0.9)
+        }}
+      >
+        <ArrowForwardIcon />
+      </Button>
+    </>
+  )
+}
